Refactored API, API now found in smallengine/se_*.h files
This code started off as a static library, however moved to a dynamic library.
The API has been refactored to reflect this, with each module (each sub dir of
the src folder) receiving it's own header file containing all struct defs and
function prototypes. This is included using

ie, to include the graphics module:

The code in the src folders has been changed to reflect this, as well as the
cmake files.