Test sources updated to use new API
The test files currently need to be compiled against the library after
it has been installed. This can be done with the following command,
assuming the current directory is the folder containing the test source
and the test being compiled is named 'test.c':

gcc test.c -o test -lsmallengine -lm -lSDL2

The math and SDL2 libs may only be necessary for some test files. This
is a temporary solution until a more suitable test framework such as
gtest is implemented.